Reworked rsu-status-checker and postgres writes along with unit tests and postgres script updates
