+++ docker-compose-addons.yml
@@ -52,7 +52,9 @@ services:
     environment:
       RSU_PING: ${RSU_PING}
       ZABBIX: ${ZABBIX}
-      RSU_SNMP_FETCH: ${RSU_SNMP_FETCH}
+      RSU_MSGFWD_FETCH: ${RSU_MSGFWD_FETCH}
+      RSU_SECURITY_FETCH: ${RSU_SECURITY_FETCH}
+      RSU_HEALTH_FETCH: ${RSU_HEALTH_FETCH}
 
       ZABBIX_ENDPOINT: ${ZABBIX_ENDPOINT}
       ZABBIX_USER: ${ZABBIX_USER}
