+++ resources/sql_scripts/CVManager_CreateTables.sql
@@ -353,6 +353,27 @@ CREATE TABLE IF NOT EXISTS public.scms_health
 		ON DELETE NO ACTION
 );
 
+-- Create rsu_health table
+CREATE SEQUENCE public.rsu_health_rsu_health_id_seq
+   INCREMENT 1
+   START 1
+   MINVALUE 1
+   MAXVALUE 2147483647
+   CACHE 1;
+
+CREATE TABLE IF NOT EXISTS public.rsu_health
+(
+   rsu_health_id integer NOT NULL DEFAULT nextval('rsu_health_rsu_health_id_seq'::regclass),
+   timestamp timestamp without time zone NOT NULL,
+   health integer NOT NULL,
+   rsu_id integer NOT NULL,
+   CONSTRAINT rsu_health_pkey PRIMARY KEY (rsu_health_id),
+   CONSTRAINT fk_rsu_id FOREIGN KEY (rsu_id)
+		REFERENCES public.rsus (rsu_id) MATCH SIMPLE
+		ON UPDATE NO ACTION
+		ON DELETE NO ACTION
+);
+
 -- Create snmp_msgfwd_type table
 CREATE SEQUENCE public.snmp_msgfwd_type_id_seq
    INCREMENT 1
