+++ sample.env
@@ -331,7 +331,9 @@ RSU_PING=True
 ZABBIX=False
 
 # Fetches SNMP configuration data for all RSUs
-RSU_SNMP_FETCH=True
+RSU_MSGFWD_FETCH=True
+RSU_SECURITY_FETCH=True
+RSU_HEALTH_FETCH=True
 
 # Zabbix endpoint and API authentication
 # Only used when ZABBIX is 'True'
